+++ client/src/components/Overview/OverviewWidget.jsx
@@ -1,5 +1,5 @@
-import SellerTable from './SellerTable.jsx';
-import FormWidget from './FormWidget.jsx';
+import SellerTable from '../Seller/SellerTable.jsx';
+import FormWidget from '../Form/FormWidget.jsx';
 const React = require('react');
 
 class OverviewWidget extends React.Component {
