+++ client/src/components/Stock/Stock.jsx
