+++ client/src/components/Stock/Stock.test.js
@@ -1,0 +1,1 @@
+
