+++ package.json
@@ -21,8 +21,10 @@
   "homepage": "https://github.com/Zheng-Yi-Sao/ProductOverview#readme",
   "devDependencies": {
     "@babel/core": "^7.14.0",
+    "@babel/plugin-syntax-jsx": "^7.12.13",
     "@babel/preset-env": "^7.14.1",
     "@babel/preset-react": "^7.13.13",
+    "@wojtekmaj/enzyme-adapter-react-17": "^0.6.1",
     "babel-loader": "^8.2.2",
     "css-loader": "^5.2.4",
     "jest": "^26.6.3",
@@ -37,7 +39,6 @@
     "body-parser": "^1.19.0",
     "express": "^4.17.1",
     "faker": "^5.5.3",
-    "jest-enzyme": "^7.1.2",
     "jquery": "^3.6.0",
     "moment": "^2.29.1",
     "mongoose": "^5.12.7",
@@ -45,6 +46,7 @@
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
     "react-test-renderer": "^17.0.2",
+    "save": "^2.4.0",
     "uuid": "^8.3.2"
   }
 }
